% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{adv_clones}
\alias{adv_clones}
\title{adv_clones}
\format{
A data frame with 111 rows and 6 variables:
\describe{
\item{ sort }{ row number }
\item{ current_clone_id }{ current clone ID per CIP naming convention }
\item{ old_family_code }{clonal family code per previous naming convention }
\item{ year_of_cross }{ year of cross }
\item{female }{ female parent }
\item{male }{ male parent }
}
}
\source{
CIP
}
\usage{
adv_clones
}
\description{
data containing advanced clones under consideration.
}
\keyword{datasets}
