% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_row.R
\name{assign_benches}
\alias{assign_benches}
\title{assign clones to tissue culture benches}
\usage{
assign_benches(x, id_col = "row_id")
}
\arguments{
\item{x}{dataframe}

\item{id_col}{row id index variable}
}
\value{
dataframe
}
\description{
assign clones to tissue culture benches
}
