% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_location.R
\name{capture_location}
\alias{capture_location}
\title{use \code{purrr::imap} to capture location of a trial from trial file
ipurrr::map applies a function to each element of a vector and its index}
\usage{
capture_location(lst)
}
\arguments{
\item{lst}{a list object}
}
\value{
a vector of same length as \code{lst}
}
\description{
use \code{purrr::imap} to capture location of a trial from trial file
ipurrr::map applies a function to each element of a vector and its index
}
