% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_chars.R
\name{check_gene_chars}
\alias{check_gene_chars}
\title{check length of accession family name - should be 9}
\usage{
check_gene_chars(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a data frame
}
\description{
check length of accession family name - should be 9
}
