% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{check_vars}
\alias{check_vars}
\title{find variables not collecte but should have been}
\usage{
check_vars(x, y = c("npe", "nph", "ntp"))
}
\arguments{
\item{x}{a data frame}

\item{y}{a vector of variables to find}
}
\value{
a list
}
\description{
find variables not collecte but should have been
}
