% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_clone_name.R
\name{clean_clone_name}
\alias{clean_clone_name}
\title{Clean accession names}
\usage{
clean_clone_name(x)
}
\arguments{
\item{x}{a character vector of clone/genotype IDs}
}
\value{
a vector of same length as \code{x}
}
\description{
Clean accession names
}
