% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file_names.R
\name{clean_dir_name}
\alias{clean_dir_name}
\title{clean file name using \code{gsub}}
\usage{
clean_dir_name(x)
}
\arguments{
\item{x}{a character vector of file directory or file name}
}
\value{
a character vector
}
\description{
clean file name using \code{gsub}
}
\examples{

f <- system.file("uon-trial-1.csv", package = "pbwrangler")
clean_dir_name(f)
}
