% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file_names.R
\name{clean_dir_name_c}
\alias{clean_dir_name_c}
\title{clean file name using \code{gsub}}
\usage{
clean_dir_name_c(x)
}
\arguments{
\item{x}{a character vector of file directory or file name}
}
\value{
a character vector
}
\description{
clean file name using \code{gsub}
}
