% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_file_names.R
\name{clean_file_names}
\alias{clean_file_names}
\title{clean file names using \code{gsub}}
\usage{
clean_file_names(out)
}
\arguments{
\item{out}{a list of trial dataframes}
}
\value{
a list with names cleaned
}
\description{
clean file names using \code{gsub}
}
