% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{clean_var_names}
\alias{clean_var_names}
\title{clean variable names}
\usage{
clean_var_names(x)
}
\arguments{
\item{x}{a character vector of variable names}
}
\value{
a character vector
}
\description{
clean variable names
}
