% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{combine_meta_files}
\alias{combine_meta_files}
\title{read meta-files from separate trials in a season into 1 file}
\usage{
combine_meta_files(season, path = out_dir)
}
\arguments{
\item{season}{trial season such as \code{"season-2024"}}

\item{path}{directory containing meta files}
}
\value{
a dataframe
}
\description{
read meta-files from separate trials in a season into 1 file
}
\examples{

list_files(system.file(package = "pbwrangler"))[1:5]
}
