% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_df.R
\name{compare_df}
\alias{compare_df}
\title{Compare field experiment data to see if they share designs, variables, etc.}
\usage{
compare_df(x, y)
}
\arguments{
\item{x}{first dataframe}

\item{y}{second dataframe}
}
\value{
a dataframe with difference/similarities btwn \code{x}, \code{y}
}
\description{
Compare field experiment data to see if they share designs, variables, etc.
}
