% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{compute_cols}
\alias{compute_cols}
\title{Compute nmtp, mtwp (no/marketable tuber weight per plot)}
\usage{
compute_cols(x)
}
\arguments{
\item{x}{a list of dataframes}
}
\value{
a list
}
\description{
Compute nmtp, mtwp (no/marketable tuber weight per plot)
}
