% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{convert_to_character}
\alias{convert_to_character}
\title{Coerce columns to character}
\usage{
convert_to_character(x, to_char = c("obs", "notes"))
}
\arguments{
\item{x}{a dataframe}

\item{to_char}{column names to coerce to character}
}
\value{
a dataframe
}
\description{
Coerce columns to character
}
