% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{convert_to_factor}
\alias{convert_to_factor}
\title{Coerce columns to factors}
\usage{
convert_to_factor(
  x,
  to_fact = c("rep", "plot", "loc", "block", "trial", "row", "col", "type", "check",
    "clone", "geno")
)
}
\arguments{
\item{x}{a dataframe}

\item{to_fact}{a vector of column names to coerce to factor}
}
\value{
a dataframe
}
\description{
Coerce columns to factors
}
