% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{onvert column values that numeric but read as character to numeric}
\usage{
convert_to_numeric(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
onvert column values that numeric but read as character to numeric
}
