% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family_vars.R
\name{create_family_vars}
\alias{create_family_vars}
\title{create new family code per new naming convention}
\usage{
create_family_vars(x)
}
\arguments{
\item{x}{a dataframe of trial data}
}
\value{
a dataframe
}
\description{
create new family code per new naming convention
}
