% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_file.R
\name{create_meta_file}
\alias{create_meta_file}
\title{create a meta-data file with design factors, planting dates etc}
\usage{
create_meta_file(x, season, d_dir = out_dir)
}
\arguments{
\item{x}{a list of trials to generate metadata for}

\item{season}{trial season e.g \code{"season-2024"}}

\item{d_dir}{folder directory to write the meta-data e.g., \code{"C://Users/Documents"}}
}
\description{
create a meta-data file with design factors, planting dates etc
}
