% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family_vars.R
\name{create_pedigree}
\alias{create_pedigree}
\title{create pedigree (ancestry)}
\usage{
create_pedigree(x)
}
\arguments{
\item{x}{a dataframe of genotypes created by \code{create_family_vars()}}
}
\value{
a dataframe
}
\description{
create pedigree (ancestry)
}
\seealso{
\code{\link[=create_family_vars]{create_family_vars()}}
}
