% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{create_td}
\alias{create_td}
\title{A wrapper for \code{statgenSTA::createTD} for a \code{row-col} design model fitting}
\usage{
create_td(x, design)
}
\arguments{
\item{x}{datframe of trial data}

\item{design}{character vector of experimental design e.g \code{"rowcol"}, \code{"rcbd"}}
}
\value{
a list object of class TD
}
\description{
A wrapper for \code{statgenSTA::createTD} for a \code{row-col} design model fitting
}
