% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{drop_all_data}
\alias{drop_all_data}
\title{drop dataframe objects from a list of dataframes}
\usage{
drop_all_data(x, pattern = "^all|a1_table")
}
\arguments{
\item{x}{a list of dataframes}

\item{pattern}{pattern a character vector containing regular expression}
}
\value{
a list of dataframes
}
\description{
drop dataframe objects from a list of dataframes
}
