% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{drop_empty_dfs}
\alias{drop_empty_dfs}
\title{drop from a list dataframes with no dimension or  columns having all \code{NA}}
\usage{
drop_empty_dfs(x)
}
\arguments{
\item{x}{a list of dataframes}
}
\value{
a list
}
\description{
drop from a list dataframes with no dimension or  columns having all \code{NA}
}
