% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{drop_nas}
\alias{drop_nas}
\title{delete columns with all values missing(\code{NA})}
\usage{
drop_nas(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
A dataframe without completely missing columns
}
\description{
delete columns with all values missing(\code{NA})
}
