% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{drop_null_dfs}
\alias{drop_null_dfs}
\title{drop null dataframes from a list of dataframes}
\usage{
drop_null_dfs(x)
}
\arguments{
\item{x}{a list of dataframes}
}
\value{
a list
}
\description{
drop null dataframes from a list of dataframes
}
