% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{drop_sheets}
\alias{drop_sheets}
\title{drop workbook sheets not of interest}
\usage{
drop_sheets(x, pattern = "min|pivot|sheet|weather|soil|all|unse|sele")
}
\arguments{
\item{x}{a list of excel sheets}

\item{pattern}{patern to exclude}
}
\value{
a character vector
}
\description{
drop workbook sheets not of interest
}
