% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{drop_zeros}
\alias{drop_zeros}
\title{delete columns with all values 0}
\usage{
drop_zeros(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe with columns with all values zero deleted
}
\description{
delete columns with all values 0
}
