% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_row.R
\name{duplicate_row}
\alias{duplicate_row}
\title{duplicate a row}
\usage{
duplicate_row(x, n = 8)
}
\arguments{
\item{x}{a dataframe}

\item{n}{number of copies to create}
}
\value{
a dataframe
}
\description{
duplicate a row
}
