% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{extract_blups}
\alias{extract_blups}
\title{extract BLUPs from a \code{REML} model}
\usage{
extract_blups(x, pred)
}
\arguments{
\item{x}{an object of class \code{STA} from \code{fit_td}}

\item{pred}{a character vector of predictors: \verb{c("BLUEs", "seBLUEs)} or \verb{c("BLUPs", "seBLUPs)}}
}
\value{
a dataframe
}
\description{
extract BLUPs from a \code{REML} model
}
