% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{family_code}
\alias{family_code}
\title{country_level_data}
\format{
A data frame with 222 rows and 6 variables:
\describe{
\item{ sort }{ row number }
\item{ new_family_code }{ new family code per CIP new naming convention }
\item{ old_family_code }{clonal family code per previous naming convention }
\item{ year_of_cross }{ year of cross }
\item{female }{ female parent }
\item{male }{ male parent }
}
}
\source{
Observatoire de l'Habitat
}
\usage{
family_code
}
\description{
data in 'luHousePrices'.
}
\keyword{datasets}
