% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{find_var}
\alias{find_var}
\title{find variables in df}
\usage{
find_var(x, var)
}
\arguments{
\item{x}{dataframe}

\item{var}{column name}
}
\value{
a character vector
}
\description{
find variables in df
}
