% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{fit_td}
\alias{fit_td}
\title{A wrapper for \code{statgenSTA::fitTD} for REML analysis}
\usage{
fit_td(x, trait, spatial = FALSE)
}
\arguments{
\item{x}{TD object generated by \code{create_td}}

\item{trait}{a character vector phenotypic traits to model e.g \code{ttyna}}

\item{spatial}{logical, whether to fit spatial structure}
}
\value{
an object of class STA
}
\description{
A wrapper for \code{statgenSTA::fitTD} for REML analysis
}
\seealso{
\code{\link[=create_td]{create_td()}}
}
