% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_location.R
\name{gen_familycode}
\alias{gen_familycode}
\title{a regular expr deleting everything after last fullstop(\code{.}) or last underscore(\verb{_})}
\usage{
gen_familycode(x)
}
\arguments{
\item{x}{a character vector of accession names}
}
\value{
a character vector
}
\description{
a regular expr deleting everything after last fullstop(\code{.}) or last underscore(\verb{_})
}
