% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_location.R
\name{gen_location}
\alias{gen_location}
\title{generate possible location from file paths}
\usage{
gen_location(x)
}
\arguments{
\item{x}{a character vector of file directory}
}
\value{
a character vector
}
\description{
generate possible location from file paths
}
