% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{gen_row_col}
\alias{gen_row_col}
\title{generate rows and columns after \code{split_by_chunk}}
\usage{
gen_row_col(x)
}
\arguments{
\item{x}{a list of dataframes}
}
\value{
a list
}
\description{
generate rows and columns after \code{split_by_chunk}
}
