% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_row.R
\name{gen_uniqueid}
\alias{gen_uniqueid}
\title{generate unique ids}
\usage{
gen_uniqueid(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a character vector of hashed ids
}
\description{
generate unique ids
}
