% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_prep.R
\name{geno_by_tubers}
\alias{geno_by_tubers}
\title{Distribute clones to match the available seeds during replication}
\usage{
geno_by_tubers(df)
}
\arguments{
\item{df}{a dataframe of available seed}
}
\value{
a list
}
\description{
Distribute clones to match the available seeds during replication
}
\examples{

data("ilri", package = "pbwrangler")
ins_ilri <- geno_by_tubers(ilri)
lapply(ins_ilri, head, 5)
}
