% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_invalid_cols.R
\name{get_design_factors}
\alias{get_design_factors}
\title{Get experimental design factors from a trial dataframe/fieldbook}
\usage{
get_design_factors()
}
\value{
a character vector
}
\description{
Get experimental design factors from a trial dataframe/fieldbook
}
