% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{get_fieldbooks}
\alias{get_fieldbooks}
\title{get field trial fieldbooks in a particular folder}
\usage{
get_fieldbooks(path = t_dir, season, sub_folder = NULL, s = ".csv|.xlsx")
}
\arguments{
\item{path}{character string specifying the directory containing the fieldbooks}

\item{season}{character vector of trial season}

\item{sub_folder}{character vector of sub-directory. defaults to NULL}

\item{s}{character grep pattern of file extensions to return}
}
\value{
a dataframe
}
\description{
get field trial fieldbooks in a particular folder
}
