% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_invalid_cols.R
\name{get_valid_columns}
\alias{get_valid_columns}
\title{get column names defined in ontology}
\usage{
get_valid_columns(x, crop = "pt")
}
\arguments{
\item{x}{a dataframe}

\item{crop}{one of "pt" or "sp"}
}
\value{
a dataframe
}
\description{
get column names defined in ontology
}
