% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{\%nin\%}
\alias{\%nin\%}
\title{\verb{\%nin\%} returns a logical vector if there is a match or not for left operand}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{a vector}

\item{table}{a vector matching the mode of \code{x}}
}
\value{
logical vector
}
\description{
\verb{\%nin\%} returns a logical vector if there is a match or not for left operand
}
