% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{ilri}
\alias{ilri}
\title{ilri}
\format{
A data frame with 55 row and 2 variables:
\describe{
\item{ geno }{ genotype }
\item{ number_of_tubers }{ number of tubers}
}
}
\source{
CIP
}
\usage{
ilri
}
\description{
data containing advanced clones under consideration.
}
\keyword{datasets}
