% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{join_by_keys}
\alias{join_by_keys}
\title{Join predictor and heritability (\code{h2}) data objects to one dataframe}
\usage{
join_by_keys(x, y)
}
\arguments{
\item{x}{first data object either predictors or \code{h2}}

\item{y}{second data object}
}
\value{
a dataframe
}
\description{
Join predictor and heritability (\code{h2}) data objects to one dataframe
}
