% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{join_data_meta}
\alias{join_data_meta}
\title{join data and meta data}
\usage{
join_data_meta(data_lst, meta_lst)
}
\arguments{
\item{data_lst}{a list of trial dfs}

\item{meta_lst}{a list of accompanying meta data}
}
\value{
a list
}
\description{
join data and meta data
}
