% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_design_files}
\alias{list_design_files}
\title{List design files after designing trial experiments in R}
\usage{
list_design_files(
  dir = rand_dir,
  season = "season-2024-2025",
  s = ".csv|.xlsx",
  sub_dir = NULL,
  ...
)
}
\arguments{
\item{dir}{directory containing the fieldboooks}

\item{season}{trial seasom}

\item{s}{a vector of charcter string to grep e.g \code{s = ".csv|.xlsx"}}

\item{sub_dir}{sub-directory if filedbooks are nested in a sub-directory}

\item{...}{any other argument}
}
\value{
a list of filedbooks
}
\description{
List design files after designing trial experiments in R
}
