% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_files.R
\name{list_files}
\alias{list_files}
\title{List file in a directory}
\usage{
list_files(folder, subset = FALSE, n = NULL, s = ".csv|.xlsx", ...)
}
\arguments{
\item{folder}{directory with files e.g. "C://users/user/x}

\item{subset}{a logical value whether or not to return a subset of files in \code{folder}}

\item{n}{integer. number of files to return if subsetting}

\item{s}{character. pattern to look for e.g., s = ".csv|.xlsx"}

\item{...}{any other parameters to specify}
}
\value{
a character vector of listed files
}
\description{
List file in a directory
}
