% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_note_obs.R
\name{merge_note_obs}
\alias{merge_note_obs}
\title{Merge notes & obs to one column}
\usage{
merge_note_obs(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
Merge notes & obs to one column
}
