% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/included-datasets.R
\docType{data}
\name{meta}
\alias{meta}
\title{adv_clones}
\format{
A data frame with 1 row and 17 variables:
\describe{
\item{ trial_name }{ trial name }
\item{ breeding_program }{ breeding program }
\item{ location }{trial location }
\item{ year }{ trial year }
\item{transplanting_date }{ date of transplanting }
\item{design_type }{ type of field design }
\item{ description }{ A description field detailing the trial }
\item{trial_type }{ trial type }
\item{ plot_width }{ plot width }
\item{ plot_length }{plot length }
\item{ planting_date }{ planting date }
\item{harvest_date }{ harvest date }
\item{number_of_plants_per_ridge }{ number of plants per ridge }
\item{ number_of_ridges_per_plot }{ number of ridges per plot }
\item{ space_between_ridges }{ space between ridges }
\item{ space_between_plants_in_ridges }{space between plants in ridges }
\item{ number_of_plants_per_plot }{ number of plants per plot }
}
}
\source{
CIP
}
\usage{
meta
}
\description{
data containing advanced clones under consideration.
}
\keyword{datasets}
