% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{my_left}
\alias{my_left}
\title{Extract first n elements of a string}
\usage{
my_left(x, num)
}
\arguments{
\item{x}{a character vector of any length}

\item{num}{an integer - the number of elements to extract}
}
\value{
a character vector
}
\description{
Extract first n elements of a string
}
