% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{my_theme}
\alias{my_theme}
\title{plot weather data coming from HOBO data loggers}
\usage{
my_theme(
  plot,
  caption,
  title,
  legend_pos = "off",
  color_c = c("#003f5c", "#58508d", "#bc5090", "#ff6361", "#ffa600"),
  color_bg = "grey99",
  color_axis = "#273B50",
  color_title = "#273B50",
  color_panelbg = "#F6F2EE",
  b_font = "Source Serif Pro"
)
}
\arguments{
\item{plot}{plot to theme}

\item{caption}{plot caption}

\item{title}{plot title}

\item{legend_pos}{a character vector specifying legend position}

\item{color_c}{a character vector of hex color codes}

\item{color_bg}{a character vector of hex color codes}

\item{color_axis}{a character vector of hex color codes}

\item{color_title}{a character vector of hex color codes}

\item{color_panelbg}{a character vector of hex color codes}

\item{b_font}{a character vector of font family}
}
\description{
plot weather data coming from HOBO data loggers
}
