% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbwrangler-package.R
\docType{package}
\name{pbwrangler-package}
\alias{pbwrangler}
\alias{pbwrangler-package}
\title{pbwrangler: Potato Breeding Phenotypic Data Wrangler}
\description{
`pbwrangler` is a collection of in-house functions used in curating potato breeding data for analysis and archiving into breeding databases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bokola/pbwrangler}
  \item Report bugs at \url{https://github.com/Bokola/pbwrangler/issues}
}

}
\author{
\strong{Maintainer}: Basil Okola \email{okolabasilowiti@gmail.com}

}
\keyword{internal}
