% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{plot_weather}
\alias{plot_weather}
\title{plot weather data coming from HOBO data loggers}
\usage{
plot_weather(
  df = df_w_summ,
  var = "temperature_mean",
  color_c = c("#003f5c", "#58508d", "#bc5090", "#ff6361", "#ffa600"),
  color_bg = "grey99",
  color_axis = "#273B50",
  color_title = "#273B50",
  color_panelbg = "#F6F2EE",
  b_font = "Source Serif Pro"
)
}
\arguments{
\item{df}{a daframe}

\item{var}{a character vector variable}

\item{color_c}{a character vector of hex color codes}

\item{color_bg}{a character vector of hex color codes}

\item{color_axis}{a character vector of hex color codes}

\item{color_title}{a character vector of hex color codes}

\item{color_panelbg}{a character vector of hex color codes}

\item{b_font}{a character vector of font family}
}
\value{
a plot
}
\description{
plot weather data coming from HOBO data loggers
}
