% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_prep.R
\name{rand_Prep}
\alias{rand_Prep}
\title{Partially replicated field design}
\usage{
rand_Prep(
  tot,
  ins,
  rowD,
  trial,
  n_dummies = 0,
  loc = NULL,
  totReps,
  trtrepP,
  trtgroup,
  block_lst,
  season = "season-2024-2025",
  path = t_dir,
  check = c("Shangi", "Unica", "Sagitta", "Sherekea"),
  dummy = c("Unica", "Shangi"),
  to_add = 4
)
}
\arguments{
\item{tot}{integer. total number of unique clones/genotypes to be randomized to field}

\item{ins}{a character vector of genotypes}

\item{rowD}{integer. number of rows in the field}

\item{trial}{character. trial name}

\item{n_dummies}{integer. number of dummies to complete a rectangular layout}

\item{loc}{character. trial location}

\item{totReps}{integer. total number of plots: row by col}

\item{trtrepP}{numeric vector. replications of \code{ins} given in the form
\verb{rep(c(vector of reps), c(vector of number of clones))} e.g.,
\code{rep(c(1,8), c(304, 4))}}

\item{trtgroup}{numeric vector. replication of treatment groups. samilar repliction as \code{trtrepP}}

\item{block_lst}{a list specifying blocking of the field}

\item{season}{season of trial}

\item{path}{character specifying path to write the design}

\item{check}{a character vector of checks to fill rectangular grid}

\item{dummy}{a character vector of dummy checks to fill rectangular grid}

\item{to_add}{integer. number of checks to add to complete the rectangular grid}
}
\description{
Partially replicated field design
}
