% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_row_col.R
\name{randomize_res_row_col}
\alias{randomize_res_row_col}
\title{resolvable row column design: plots = row by col; test clones replicated once, checks more than once}
\usage{
randomize_res_row_col(
  clones,
  tot,
  trial,
  totReps,
  trtrepP,
  block_lst,
  rowD,
  n_dummies,
  season,
  path = rand_dir,
  check = c("Shangi", "Unica", "Sagitta", "Sherekea"),
  dummy = c("Unica", "Shangi"),
  to_add = 4
)
}
\arguments{
\item{clones}{a dataframe with \code{geno} column}

\item{tot}{integer. total number of unique clones/genotypes to be randomized to field}

\item{trial}{character. trial}

\item{totReps}{integer. total number of plots: row by col}

\item{trtrepP}{numeric vector. replications of \code{ins}}

\item{block_lst}{a list specifying blocking of the field}

\item{rowD}{integer. number of rows in the field}

\item{n_dummies}{integer. number of dummies to complete a rectangular layout}

\item{season}{season of trial}

\item{path}{character specifying path to write the design}

\item{check}{a character vector of checks to fill rectangular grid}

\item{dummy}{a character vector of dummy checks to fill rectangular grid}

\item{to_add}{integer. number of checks to add to complete the rectangular grid}
}
\description{
resolvable row column design: plots = row by col; test clones replicated once, checks more than once
}
