% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_accessions.R
\name{read_accessions}
\alias{read_accessions}
\title{read accession files that have been fixed and confirmed correct}
\usage{
read_accessions(dir = out_dir)
}
\arguments{
\item{dir}{a character vector of path to accession files}
}
\value{
a dataframe of fixed accession names
}
\description{
read accession files that have been fixed and confirmed correct
}
