% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_workbooks.R
\name{read_workbooks}
\alias{read_workbooks}
\title{A wrapper for \code{readr}'s \code{read_csv} & \code{readxl}'s \code{read_excel} for reading trial data}
\usage{
read_workbooks(
  dir = t_dir,
  season = "Season 2023",
  read_subfolder = FALSE,
  s = ".csv|.xlsx",
  sub_folder = "Data/Curated Data",
  subset = FALSE,
  n = NULL,
  multiple = FALSE,
  sheet_name = NULL,
  skip = 0
)
}
\arguments{
\item{dir}{trial data home directory}

\item{season}{trial season}

\item{read_subfolder}{logical. \code{TRUE} if data lives inside a subfolder in \code{season}}

\item{s}{a regular expression used to keep relevant files}

\item{sub_folder}{character of subfolder containing data e.g \code{"Data/Curated Data"}}

\item{subset}{logical. \code{TRUE} if you want to read just a subset of files}

\item{n}{integer. Number of files to subset e.g, \code{1}, \code{-1} \code{1:4}.
negative prefix if you want to exclude a file at a specific index}

\item{multiple}{logical. \code{TRUE} if you want to read nultiple excel sheets from the same workbook}

\item{sheet_name}{character. used to specify excel sheet name of interest, e.g., \code{"fieldbook"}}

\item{skip}{integer. number of rows to skip when reading data from an excel workbook}
}
\value{
a list of trial dataframes
}
\description{
A wrapper for \code{readr}'s \code{read_csv} & \code{readxl}'s \code{read_excel} for reading trial data
}
