% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{recode_var}
\alias{recode_var}
\title{Coarce variable to numeric or character}
\usage{
recode_var(
  x,
  to_num = c("plot", "tuber_size_mm", "rep", "row", "col", "yield"),
  to_char = c("flowering_date|obs|^note")
)
}
\arguments{
\item{x}{a dataframe}

\item{to_num}{columns to coerce to numeric}

\item{to_char}{columns to coerce to character}
}
\value{
a dataframe
}
\description{
Coarce variable to numeric or character
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f) \%>\% `[[`(1)
df_c <- recode_var(df)
}
