% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks}
\alias{run_checks}
\title{run data quality checks using \code{{st4gi}} functions}
\usage{
run_checks(x, sz = 10000, btwn = 0.75, within = 0.3, ...)
}
\arguments{
\item{x}{a dataframe}

\item{sz}{size of farm, here taken to be 10000 sq metres}

\item{btwn}{spacing between ridges in metres}

\item{within}{spacing between plants in the same ridge in metres}

\item{...}{additional arguments}
}
\value{
a dataframe
}
\description{
run data quality checks using \code{{st4gi}} functions
}
