% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_trials_.R
\name{run_data_processes}
\alias{run_data_processes}
\title{compute derived phenotypic variables}
\usage{
run_data_processes(x, sz = 10000, btwn = 0.75, within = 0.3, crop = "pt")
}
\arguments{
\item{x}{a dataframe}

\item{sz}{size of farm, here taken to be 10000 sq metres}

\item{btwn}{spacing between ridges in metres}

\item{within}{spacing between plants in the same ridge in metres}

\item{crop}{one of \code{'pt'} or \code{'sp'}}
}
\value{
a dataframe
}
\description{
compute derived phenotypic variables
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f)  \%>\%
  purrr::map(
    .,
    ~ dplyr::mutate(
      .,
      space_between_ridges = "0.75m",
      space_between_plants_in_ridges = "0.3m",
      number_of_plants_per_plot = 10
    )
  )

df_out <- pre_process_trials(df) |> process_trials() \%>\%
  purrr::map(., run_data_processes)
purrr::map(df_out, names_df)
}
