% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_var.R
\name{select_cols}
\alias{select_cols}
\title{A wrapper for \code{dplyr::select} function}
\usage{
select_cols(x, cols)
}
\arguments{
\item{x}{a dataframe}

\item{cols}{a character vector of column names}
}
\value{
a dataframe
}
\description{
A wrapper for \code{dplyr::select} function
}
