% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{split_by_chunk}
\alias{split_by_chunk}
\title{split a dataframe into a list by row numbe}
\usage{
split_by_chunk(x, chunk)
}
\arguments{
\item{x}{a dataframe}

\item{chunk}{integer. rows for each list element}
}
\description{
split a dataframe into a list by row numbe
}
