% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_invalid_cols.R
\name{subset_invalid_cols}
\alias{subset_invalid_cols}
\title{get invalid names (labels not in ontology) & add geno}
\usage{
subset_invalid_cols(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
get invalid names (labels not in ontology) & add geno
}
