% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{sum_rowwise}
\alias{sum_rowwise}
\title{dplyr::rowwise sum target columns}
\usage{
sum_rowwise(x, target_cols)
}
\arguments{
\item{x}{a dataframe}

\item{target_cols}{character vector of column names}
}
\value{
a dataframe
}
\description{
dplyr::rowwise sum target columns
}
