% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_prep.R
\name{trial_design_meta}
\alias{trial_design_meta}
\title{generate trial design inputs i.e., geno replications and blocking list}
\usage{
trial_design_meta(
  trep = list(c(3, 2, 1, 5), c(6, 11, 36, 4)),
  trgroup = trep,
  block_list = list(c(6, 4), c(6, 2))
)
}
\arguments{
\item{trep}{numeric vector of treatment/genotype replications}

\item{trgroup}{numeric vector of treatment group replication}

\item{block_list}{a list of blocking field layout}
}
\value{
a list
}
\description{
generate trial design inputs i.e., geno replications and blocking list
}
\examples{
d <- tempdir()
data("ilri", package = "pbwrangler")
ins_ilri <- geno_by_tubers(ilri)
ilri_prep <- rand_Prep(
  tot = 57,
  ins = ins_ilri,
  rowD = 12,
  trial = "KE24ILR-BIO-IT01",
  n_dummies = 5,
  loc = "ilri",
  totReps =96,
  trtrepP = trial_design_meta()$trep,
  trtgroup = trial_design_meta()$trgroup,
  block_lst = trial_design_meta()$block_list,
  path = NULL
)
head(ilri_prep$design)
}
