% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_geno.R
\name{update_geno}
\alias{update_geno}
\title{Update accession names to conform to naming convention}
\usage{
update_geno(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
Update accession names to conform to naming convention
}
