% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{write processed trial data to a directory}
\usage{
write_data(dir = b_out_dir, data_list, season)
}
\arguments{
\item{dir}{destination directory/folder to write to}

\item{data_list}{list of processed trial data}

\item{season}{trial season}
}
\description{
write processed trial data to a directory
}
