% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_outliers}
\alias{write_outliers}
\title{write out trial data}
\usage{
write_outliers(x, season, outliers = TRUE, dir = t_dir)
}
\arguments{
\item{x}{list of trial data}

\item{season}{trial season}

\item{outliers}{logical: select clones with outlying yield}

\item{dir}{output directory}
}
\description{
write out trial data
}
