% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_season_data}
\alias{write_season_data}
\title{Write out data for pre-processed experimental data by calling \code{write_trials()}}
\usage{
write_season_data(list_df, season)
}
\arguments{
\item{list_df}{a list of data frames}

\item{season}{seasson of experiments}
}
\description{
Write out data for pre-processed experimental data by calling \code{write_trials()}
}
\seealso{
\code{\link[=write_trials]{write_trials()}}
}
