% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_trials}
\alias{write_trials}
\title{write out trial data}
\usage{
write_trials(x, season, is_invalid = FALSE, dir = t_dir)
}
\arguments{
\item{x}{list of trial data}

\item{season}{trial season}

\item{is_invalid}{logical: select invalid columns not defined in crop ontology}

\item{dir}{output directory}
}
\description{
write out trial data
}
