% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gene_chars.R
\name{check_geno}
\alias{check_geno}
\title{get accession names that do not conform to required protocol}
\usage{
check_geno(x, season, dir = out_dir)
}
\arguments{
\item{x}{a list of dataframes}

\item{season}{trial season}

\item{dir}{output directory}
}
\value{
a data frame
}
\description{
get accession names that do not conform to required protocol
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f) 
df_checked <- check_geno(df)
}
