% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_family_vars.R
\name{create_family_vars}
\alias{create_family_vars}
\title{create new family code per new naming convention}
\usage{
create_family_vars(x)
}
\arguments{
\item{x}{a dataframe of trial data}
}
\value{
a dataframe
}
\description{
create new family code per new naming convention
}
\examples{
f <- system.file("uon-trial-1.xlsx", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f, sheet_name = "Sheet 1") \%>\%
  capture_location(.) \%>\% `[[`(1)

create_family_vars(df)[1:5, ]

}
