% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meta_file.R
\name{create_meta_file}
\alias{create_meta_file}
\title{create a meta-data file with design factors, planting dates etc}
\usage{
create_meta_file(x, season, d_dir = t_dir)
}
\arguments{
\item{x}{a list of trials to generate metadata for}

\item{season}{trial season e.g \code{"season-2024"}}

\item{d_dir}{folder directory to write the meta-data e.g., \code{"C://Users/Documents"}}
}
\value{
a list of dataframes
}
\description{
create a meta-data file with design factors, planting dates etc
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f)
create_meta_file(df, season = "season-2024", d_dir = tempdir())
}
