% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_invalid_cols.R
\name{get_ontology_labels}
\alias{get_ontology_labels}
\title{get ontology labels from \code{{st4gi}}}
\usage{
get_ontology_labels(x, crop = "pt")
}
\arguments{
\item{x}{a list of dataframes or a dataframe object}

\item{crop}{a character vecto - \code{pt} or \code{sp}}
}
\value{
a dataframe
}
\description{
get ontology labels from \code{{st4gi}}
}
