% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_td.R
\name{join_by_keys}
\alias{join_by_keys}
\title{Join predictor and heritability (\code{h2}) data objects to one dataframe}
\usage{
join_by_keys(x, y)
}
\arguments{
\item{x}{first data object either predictors or \code{h2}}

\item{y}{second data object}
}
\value{
a dataframe
}
\description{
Join predictor and heritability (\code{h2}) data objects to one dataframe
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f) \%>\%
  purrr::map(
    ., ~dplyr::mutate(
      ., year = "2024", loc = "UON", trial = "lbht"
    )
  )
df_out <- pre_process_trials(df) |> process_trials() \%>\%
  purrr::map(., run_data_processes) \%>\% `[[`(1) 
# trial design object
TD <- create_td(df_out, design = "rowcol")
# fit 
fit_TD <- fit_td(TD, trait = "atw")
# extract predictions

pred_TD <- extract_blups(fit_TD, pred = c("BLUEs", "seBLUEs"))

pred_TD[[1]][1:5,]

}
