% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_left.R
\name{my_left}
\alias{my_left}
\title{Extract first n elements of a string}
\usage{
my_left(x, num)
}
\arguments{
\item{x}{a character vector of any length}

\item{num}{an integer - the number of elements to extract}
}
\value{
a character vector
}
\description{
Extract first n elements of a string
}
\examples{
a <- c("asdf", "bnmkk")
my_left(a, num = 3)

a <- c("asdf", "bnmkk")
my_right(a, num = 3)

df <- data.frame(a = NA, b = c(1:3, NA), c = 0)
drop_nas(df)
drop_zeros(df)

}
