% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process_trials_.R
\name{pre_process_trials}
\alias{pre_process_trials}
\title{run a number of checks & transformations to pre-process a list of trial data}
\usage{
pre_process_trials(x)
}
\arguments{
\item{x}{a list of dataframes}
}
\value{
a list of dataframes
}
\description{
run a number of checks & transformations to pre-process a list of trial data
}
