% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_row_col.R
\name{randomize_noRep}
\alias{randomize_noRep}
\title{randomize trials for multiplication without replication}
\usage{
randomize_noRep(
  rowD = 23,
  tot = 207,
  ins = ins,
  trial = "KE25MOL-HERL-ST01",
  path = t_dir,
  season = "Season 2025",
  to_add = 4,
  rep = 4,
  check = c("Sagitta", "Unica", "Sherekea", "Shangi")
)
}
\arguments{
\item{rowD}{integer. rows in design}

\item{tot}{integer. total number of genotypes}

\item{ins}{dataframe. contains a column named \code{geno}}

\item{trial}{character. name of trial}

\item{path}{character. path to write design}

\item{season}{character. season to randomize}

\item{to_add}{integer. number of checks to add}

\item{rep}{integer. number of replications for each of the checks}

\item{check}{character. vector of checks}
}
\description{
randomize trials for multiplication without replication
}
\examples{


data("ilri")
dat <- ilri \%>\% dplyr::select(geno) 

d <- randomize_noRep(
  rowD = 11,
  tot = 55,
  ins = dat,
  path = NULL
)

head(d)

}
