% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_row_col.R
\name{randomize_row_col}
\alias{randomize_row_col}
\title{row column design: plots = row by col; equal rep for each treatment}
\usage{
randomize_row_col(
  clones,
  tot,
  loc,
  rowD,
  n_dummies,
  rep,
  season,
  path = rand_dir,
  check = c("Shangi", "Unica", "Sagitta", "Sherekea"),
  dummy = c("Unica", "Shangi"),
  to_add = 4
)
}
\arguments{
\item{clones}{ca dataframe with \code{geno} column}

\item{tot}{integer. total number of unique clones/genotypes to be randomized to field}

\item{loc}{character. trial location}

\item{rowD}{integer. number of rows in the field}

\item{n_dummies}{integer. number of dummies to complete a rectangular layout}

\item{rep}{integer. number of replication}

\item{season}{season of trial}

\item{path}{character specifying path to write the design}

\item{check}{a character vector of checks to fill rectangular grid}

\item{dummy}{a character vector of dummy checks to fill rectangular grid}

\item{to_add}{integer. number of checks to add to complete the rectangular grid}
}
\description{
row column design: plots = row by col; equal rep for each treatment
}
\examples{
df <- data.frame(geno = LETTERS[1:4])
rcD <-
  randomize_row_col(
    clones = df,
    tot = 6,
    rowD = 6,
    n_dummies = 0,
    to_add = 2,
    rep = 3,
    path = NULL
  )
head(rcD)
}
