% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomize_row_col.R
\name{randomize_row_col}
\alias{randomize_row_col}
\title{row column design: plots = row by col; equal rep for each treatment}
\usage{
randomize_row_col(
  clones,
  tot,
  trial,
  rowD,
  rowsinR,
  colsinR,
  n_dummies = 0,
  rep,
  season,
  path = t_dir,
  check = c("Shangi", "Unica", "Sagitta", "Sherekea"),
  dummy = c("Unica", "Shangi"),
  to_add = 4
)
}
\arguments{
\item{clones}{a dataframe with \code{geno} column}

\item{tot}{integer. total number of unique clones/genotypes to be randomized to field}

\item{trial}{character. trial}

\item{rowD}{integer. number of rows in the field}

\item{rowsinR}{integer. number of rows in template replicate block; for blocking}

\item{colsinR}{integer. number of columns in template replicate block; for blocking}

\item{n_dummies}{integer. number of dummies to complete a rectangular layout}

\item{rep}{integer. number of replication}

\item{season}{season of trial}

\item{path}{character specifying path to write the design}

\item{check}{a character vector of checks to fill rectangular grid}

\item{dummy}{a character vector of dummy checks to fill rectangular grid}

\item{to_add}{integer. number of checks to add to complete the rectangular grid
Note: to optimize on the row or column, remember that rep\emph{rowsinR = rowD or
rep}colsinR = colsD.}
}
\description{
row column design: plots = row by col; equal rep for each treatment
}
\examples{
df <- data.frame(geno = LETTERS[1:4])
rcD <-
  randomize_row_col(
    clones = df,
    trial = "KE24ILR-BW-ST01",
    tot = 6,
    rowD = 3,
    rowsinR = 3,
    colsinR = 2,
    n_dummies = 0,
    to_add = 2,
    rep = 3,
    path = NULL
  )
head(rcD$fieldbook)
}
