% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_cols.R
\name{rename_cols}
\alias{rename_cols}
\title{rename data columns to conform to ontology labels}
\usage{
rename_cols(x)
}
\arguments{
\item{x}{a dataframe}
}
\value{
a dataframe
}
\description{
rename data columns to conform to ontology labels
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f) \%>\%
  `[[`(1)
df <- rename_cols(df)
}
