% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_checks.R
\name{run_checks}
\alias{run_checks}
\title{run data quality checks using \code{{st4gi}} functions}
\usage{
run_checks(x, sz = 10, btwn = 0.75, within = 0.3, crop = "pt", ...)
}
\arguments{
\item{x}{a dataframe}

\item{sz}{number of plants per plot}

\item{btwn}{spacing between ridges in metres}

\item{within}{spacing between plants in the same ridge in metres}

\item{crop}{character. one of "sp" or "pt"}

\item{...}{additional arguments}
}
\value{
a dataframe
}
\description{
run data quality checks using \code{{st4gi}} functions
}
\examples{
f <- system.file("uon-trial-1.csv", package = "pbwrangler")
df <- read_workbooks(dir = NULL, file_to_read = f) \%>\%
  `[[`(1) \%>\% 
  dplyr::mutate(
          space_between_ridges = "0.75m",
      space_between_plants_in_ridges = "0.3m",
      number_of_plants_per_plot = 10
  )
df_checked <- run_checks(df)
}
