% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{write processed trial data to a directory}
\usage{
write_data(dir = b_out_dir, data_list, season)
}
\arguments{
\item{dir}{destination directory/folder to write to}

\item{data_list}{list of processed trial data}

\item{season}{trial season}
}
\description{
write processed trial data to a directory
}
\examples{
f <- system.file("uon-trial-1.xlsx", package = "pbwrangler")
f1 <- system.file("uon-trial-1.csv", package = "pbwrangler")

df <- read_workbooks(dir = NULL, file_to_read = f, sheet_name = "Sheet 1")
df1 <- read_workbooks(dir = NULL, file_to_read = f1)
dat <- c(df, df1)

write_data(dir = tempdir(), data_list = dat, season = "season-2024")
}
